Reflect changes on tests
